# Conflicts:
#	src/main/resources/Mappers/MealMapper/MealMapper.xml

+++ src/main/java/com/otd/onetoday_back/account/model/memberUpdateDto.java
@@ -8,7 +8,7 @@ import lombok.ToString;
 @Setter
 @ToString
 public class memberUpdateDto {
-    private long memberNoLogin;
+    private int memberNoLogin;
     private String memberId;
     private String email;
     private String name;
